traj_proxy/traj_proxy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyyaml==6.0.1
ray[default]==2.9.0
httpx==0.26.0
psycopg[binary,pool]==3.1.18
transformers==4.37.2
jinja2
